app/pyproject.toml
[tool.poetry]
name = "hangman"
version = "0.1.0"
description = "Hangman guessing game"
authors = ["Mazita <dev.mazita@gmail.com>"]
license = "GPL3"
readme = "README.md"
packages = [{ include = "app" }]

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
pydantic = "^2.5.2"
fastapi = "^0.104.1"
uvicorn = { extras = ["standard"], version = "^0.24.0.post1" }
python-multipart = "^0.0.6"
python-jose = { extras = ["cryptography"], version = "^3.3.0" }
passlib = { extras = ["bcrypt"], version = "^1.7.4" }
python-dotenv = "^1.0.0"
sqlmodel = "^0.0.14"
requests = "^2.31.0"
asyncpg = "^0.29.0"
pydantic-settings = "^2.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
